refactor: rename ModalStep to OnboardingStep
