+++ app/javascript/dashboard/routes/dashboard/start/components/OnboardingStep.vue
